level divided to artist-level and genre-level
