+++ js/screens/game/game.js
@@ -1,4 +1,4 @@
-import GameView from '../../views/game-view';
+/* import GameView from '../../views/game-view';
 import {stateGame} from '../../data/data';
 import displayAmountMistakes from '../../lib/displayAmountMistakes';
 import displayTimer from '../../lib/displayTimer';
@@ -13,6 +13,25 @@ game.init = () => {
   displayAmountMistakes(stateGame.lives);
 
   showLevel().init();
+};
+
+export default () => game;*/
+
+import GameView from '../../views/game-view';
+import setStateGame from '../../lib/setStateGame';
+import {stateGame} from '../../data/data';
+import displayTimer from '../../lib/displayTimer';
+import displayAmountMistakes from '../../lib/displayAmountMistakes';
+import changeLevelScreen from '../../lib/change-level-screen';
+
+const game = new GameView();
+game.init = () => {
+  setStateGame();
+  stateGame.answers.splice(0, 10);
+  displayTimer();
+  displayAmountMistakes(stateGame.lives);
+
+  changeLevelScreen();
 };
 
 export default () => game;
